requirements.txt
torch==2.1.1
torchvision==0.16.1
tqdm
ipdb
h5py==3.10.0
terminaltables
nltk
rouge-score==0.1.2
numpy==1.23.5
pytorch_lightning==1.5.10
hydra-core==1.3.2
optimum==1.14.1
sentence_transformers==2.2.2
positional_encodings==6.0.1
ffmpeg-python==0.2.0
transformers==4.35.2
accelerate==0.24.1
bitsandbytes